Move components to components root folder
